chore: Update environment variables and package dependencies
This commit updates the `.env` file to set the `HOST` variable to `turkeyhombre.com.ar`. It also removes the unused `USER` and `PASSWORD` variables from the `.env` file.

Additionally, the `package-lock.json` file is modified to remove an unnecessary dependency on `@babel/plugin-proposal-private-property-in-object`.

In the `src/app.js` file, the import statement for the `materialRoutes` module is changed to `productRoutes`.

Finally, the `material.controller.js` and `material.js` files are deleted as they are no longer needed.

This commit improves the configuration and dependencies of the project.